Refactor precision handling across modules to use global dtype references
- Introduced global precision variables in `precision.py` for consistent dtype management.
- Updated `fourier.py`, `spatials.py`, `tensors.py`, and `utils.py` to utilize global precision settings.
- Enhanced `set_precision` function to support various input types and maintain global dtype references.
- Added tests for precision settings in `test_precision.py` to ensure correct behavior across different precision configurations.

+++ src/pyhilbert/fourier.py
@@ -5,6 +5,7 @@ from multipledispatch import dispatch
 
 import numpy as np
 import torch
+from .precision import global_np_float_dtype
 
 from .spatials import Momentum, Offset
 from .hilbert import MomentumSpace, HilbertSpace, Mode
@@ -35,10 +36,16 @@ def fourier_transform(K: Tuple[Momentum, ...], R: Tuple[Offset, ...]) -> torch.T
         `exp(-2π i k·r)`.
     """
     ten_K = torch.from_numpy(  # (K, d)
-        np.stack([np.array(k.rep, dtype=np.float64).reshape(-1) for k in K], axis=0)
+        np.stack(
+            [np.array(k.rep, dtype=global_np_float_dtype).reshape(-1) for k in K],
+            axis=0,
+        )
     )
     ten_R = torch.from_numpy(  # (d, R)
-        np.stack([np.array(r.rep, dtype=np.float64).reshape(-1) for r in R], axis=1)
+        np.stack(
+            [np.array(r.rep, dtype=global_np_float_dtype).reshape(-1) for r in R],
+            axis=1,
+        )
     )
     exponent = -2j * np.pi * torch.matmul(ten_K, ten_R)  # (K, R)
     return torch.exp(exponent)  # (K, R)
